Win_error.error_message now changes behaviour depending on Sys.os_type
If Sys.os_type = "Win32" then Windows system errors will be decoded,
otherwise `Unix.error_message` behaviour is preserved. This allows
applications to use this function everywhere, safe in the knowledge
that a truly unknown Unix error won't be erroneously decoded as a
Windows system error on a non-Windows platform.

Signed-off-by: David Scott <dave@recoil.org>